Bug 418611 - Initial contribution of EScriptMonkey https://bugs.eclipse.org/bugs/show_bug.cgi?id=418611
This commit is the synchronization between the code send to CQ and the
work done in between. This code has been contributed by the two Eclipse
committer Arthur Daussy and Christian Pontesegger  which are both
eclipse E4 committer.

To see more detail look in the comment of the attached bug.

Signed-off-by: adaussy <arthur.daussy@atos.net>